mostly re-arranged some stuff
Just cleaning up and making things a little more organized.
