+++ Makefile
@@ -6,7 +6,7 @@
 #    By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/08/05 10:49:04 by dpentlan          #+#    #+#              #
-#    Updated: 2023/11/18 17:06:33 by dpentlan         ###   ########.fr        #
+#    Updated: 2023/11/18 18:25:36 by dpentlan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -14,7 +14,8 @@ CC				=	cc
 
 SRC 			=	main/main.c \
 					args/arg_parsing.c \
-					map_parsing/map_parsing.c \
+					map_parsing/map_data_load.c \
+					map_parsing/map_elements.c \
 					map_parsing/map_readin.c \
 					map_parsing/map_validation.c \
 					debug/map_debug.c \
